Reimplement hitbox test
